\name{CountryPopulation}
\alias{CountryPopulation}
\title{CountryPopulation}
\usage{
CountryPopulation(countryName)
}
\description{
Graphs the population of the country provided by parameter from 1950 to 2020
}
\examples{
CountryPopulation("Mexico")
}
\param{countryName}{Name of the country to graph}
