% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A time series of yearly popultation totals from United Nation Countries}
\format{
a data frame with 235 rows
\describe{
  \item{CountryName}{The name of the country}
  \item{Year}{The population for country in given year}
}
}
\usage{
WorldPopulation
}
\description{
A time series of yearly popultation totals from United Nation Countries
}
\keyword{datasets}
